Invalid code structure of wallet package Refactored, separate domain from serialization
