+++ pkg/wallet/account/account_test.go
@@ -1,7 +1,7 @@
 // Copyright 2017 The nem-toolchain project authors. All rights reserved.
 // Use of this source code is governed by a MIT license that can be found in the LICENSE file.
 
-package wallet
+package account
 
 import (
 	"encoding/hex"
